проверил весь функционал, решил проблемы с бд, переработал систему оценивания, но нужно доработать put http://localhost:8080/api/lessons/4/homework/6/comment?comment=Great%20work,  https://localhost:8080/api/lessons/{lessonId}/homeworks и https://localhost:8080/api/lessons/1/homework/1/grades

+++ src/main/java/com/example/controller/course/CourseController.java
@@ -15,6 +15,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
@@ -98,10 +99,19 @@ public class CourseController {
     public ResponseEntity<Void> deleteCourse(
             @PathVariable Long courseId,
             Principal principal) {
+
+        if (principal == null) {
+            System.out.println("Principal is null. User is not authenticated.");
+            throw new AccessDeniedException("User is not authenticated");
+        }
+
         String teacherUsername = principal.getName();
+        System.out.println("Authenticated user: " + teacherUsername);
+
         courseService.deleteCourse(courseId, teacherUsername);
         return ResponseEntity.noContent().build();
     }
+
 
     @GetMapping("/all")
     @PreAuthorize("hasAuthority('ROLE_STUDENT') or hasAuthority('ROLE_TEACHER') or hasAuthority('ROLE_ADMIN')")
